requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
channels==4.0.0
drf-yasg==1.21.7
daphne==4.0.0
whitenoise==6.11.0
gunicorn==21.2.0

# AWS and database support
boto3==1.34.0
django-storages==1.14.2
dj-database-url==2.1.0
